requirements.txt
Django==3.0.8
djangorestframework==3.11.0
ipython==7.16.1
django-extensions==3.0.3
drf-yasg==1.17.1
